+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 ## HEAD
 
+- 🔥 [2017-05-18] `engine/containers/mazeEngineGenerator`: update to v4.0.0.
+  - 🔥 Remove everything multiple-actors related.
+- 🔥 [2017-05-18] `engine/errors`: update to v1.2.0.
+  - 🔥 Remove actor parameter and prop.
+- 🚀 [2017-05-18] `blockly/executorGenerator`: update to v3.0.0.
+  - 🚀 Adapt to one and only one actor. Everything related to hardcoded values for specific actor, or arrays or maps related to multiple actors have been replaced for structures specifically defined for one and only one actor.
+  - `ActorsToActions` -> `Instructions`. `Map<number, Array<string>>` -> `Array<string>`.
+- 🚀 [2017-05-18] `gameMetadata`: update to v5.0.0.
+  - 🔥 Remove arrays for actors. Now everything will be defined for just one.
+  - `mazeData.accesses` -> `mazeData.access`. `Array<string>` -> `string`.
+  - `mazeData.actorsPositions` -> `mazeData.actorExitIdx`. `Array<[number, number]>` -> `number`.
 - 🐛 [2017-05-18] `engine/containers/mazeEngineGenerator`: update to v3.4.2.
   - Fix #12. Definitely! But with a better approach: it requires the new `leave_maze` action to be called for the number to leave.
 - 🚀 [2017-05-18] `engine/errors`: update to v1.1.0.